backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
torch
pandas
pickle-mixin
